requirements.txt
pycryptodome==3.18.0
bitcoinlib==0.6.5
web3==6.5.0
mnemonic==0.20
